Don't allow builds to collide.
If multiple subscriptions are triggered at once, too many events will
fire, and the build task will start building while another build task is
already running. This creates max callstack exceeded messages, so let's
wait for builds to finish before running the next one.

If this wasn't a nested promise it wouldn't be an issue, but because
every step is a promise, a new build may be started in between another
build finishing and writing out it's payload.