Fix a bug where the results of a next page query were generating invalid next page queries
This adds a test to validate that the value of `$first` on the next page
of a next page is always set to the _value_ of the previous first.
Previously, we were assuming that the value of `first` on a field was
always a literal and not a variable name. This now handles value
literals and var names.

Also standardizes on a way to check if a thing is a variable